Library of Babel
Added two functions that construct an example of the theoretical Library of Babel
A library capable of holding all the knowledge that could ever possibly be written down.
The function calculates a "page number" for a given string of any length, assuming every page has the same number of characters as the string.
The inverse function returns a string corresponding to a page number.

Cool trick, could be useful one day